move to lib/utils
