+++ lib/features/dataset/display.dart
@@ -48,8 +48,8 @@ import 'package:rattle/providers/vars/types.dart';
 import 'package:rattle/r/execute.dart';
 import 'package:rattle/r/extract.dart';
 import 'package:rattle/r/extract_vars.dart';
-import 'package:rattle/r/get_large_factors.dart';
 import 'package:rattle/utils/debug_text.dart';
+import 'package:rattle/utils/get_large_factors.dart';
 import 'package:rattle/utils/get_target.dart';
 import 'package:rattle/utils/get_unique_columns.dart';
 import 'package:rattle/utils/is_numeric.dart';
