+++ lib/utils/get_numeric.dart
@@ -29,7 +29,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'package:rattle/providers/vars/roles.dart';
 import 'package:rattle/providers/vars/types.dart';
-import 'package:rattle/r/get_large_factors.dart';
+import 'package:rattle/utils/get_large_factors.dart';
 
 /// Return a list of numeric variables that are not ignored.
